docs: presentation polish, more layouts and content
