+++ docs/training-slides/decks/intro.mdx
@@ -2,7 +2,17 @@ import ctTheme from "../src/theme.js";
 
 export const theme = ctTheme;
 
-# Documentation Types
+<TitleSlide>
+
+# Module 1: Introduction
+
+</TitleSlide>
+
+---
+
+# Documentation Types (TODO make this cards)
+
+<Horizontal>
 
 1. **Reference**:
 
@@ -11,24 +21,26 @@ export const theme = ctTheme;
 - Data field semantics - what's the domain intent of that property?
 - Interface surfaces (HTTP API surface, GraphQL schema, Java / PHP / etc)
 
-2. **Guides / Tutorials / The Book / HowTos**
+1. **Guides / Tutorials / The Book / HowTos**
 
 - Step-By Step reproducible
 - Can take side paths and alternative ways (Guide & HowTo) or not (Tutorial)
 
-3. **Product Marketing / Tech Evangelism**
+1. **Product Marketing / Tech Evangelism**
 
 - not documentation in the strict sense but important
 - short-lived, not updated or maintained
 - can be sales-y, can make temporal references (for example "our recently launched feature FooBar")
 
----
-
-## Some Terminology
+</Horizontal>
 
 ---
 
-### The Content Hierarchy
+# Some Terminology
+
+---
+
+# The Content Hierarchy
 
 1. "the Documentation"
    - everything under https://docs.commercetools.com
