+++ docs/training-slides/decks/tech-setup.mdx
@@ -2,6 +2,14 @@ import ctTheme from '../src/theme.js';
 
 export const theme = ctTheme;
 
+<TitleSlide>
+
+# Module 2: Technical Setup
+
+</TitleSlide>
+
+---
+
 # Installing Prerequisites
 
 ## Node.js
