+++ docs/training-slides/decks/writing-mdx.mdx
@@ -2,6 +2,14 @@ import ctTheme from '../src/theme.js';
 
 export const theme = ctTheme;
 
+<TitleSlide>
+
+# Module 3: Writing Markdown and MDX
+
+</TitleSlide>
+
+---
+
 # TODO Markdown Basics
 
 > TODO citation of the design goal
