+++ docs/training-slides/decks/writing-style.mdx
@@ -2,6 +2,14 @@ import ctTheme from '../src/theme.js';
 
 export const theme = ctTheme;
 
+<TitleSlide>
+
+# Module 4: Writing Style
+
+</TitleSlide>
+
+---
+
 # Why and What's a "Writing Style"
 
 ---
